requirements.txt
casbin~=1.36.1
fastapi~=0.111.0
starlette~=0.37.2
SQLAlchemy~=2.0.30
python-dotenv~=1.0.1